+++ src/qa/view/page.clj
@@ -21,7 +21,7 @@
     [:link
      {:rel "stylesheet"
       :type "text/css"
-      :href "/css/style.css"}]
+      :href "/css/styles.css"}]
     [:title "QA"]
     [:body
      [:div {:class "container"}
@@ -48,7 +48,7 @@
   (form-to {:enctype "multipart/form-data"}
            [:post "/q"]
            (anti-forgery-field)
-           (text-area "question")
+           (text-area {:id "question"} "question")
            [:br]
            (file-upload "file")
            [:br]
@@ -59,10 +59,30 @@
   [:h2 "under construction"]
   [:p "このページは q の修正画面になる。"]))
 
+;;FIXME: htmlはエスケープしなくちゃ。
+(defn ss
+ "文字列 s の n 文字以降を '...' でリプレースした文字列を返す。
+  文字列長さが n に満たない時はそのまま文字列を返す。"
+  [n s]
+  (if (< (count s) n)
+    s
+    (str (subs s 0 n) "...")))
+
+(defn st
+ "時刻表示を短くする。関数名は iso でもいいかも。
+  引数 tm は time オブジェクト。"
+  [tm]
+  (subs (str tm) 0 10)) ; hh:mm:ss を入れるなら s/10/19/
+
 (defn questions-page [qs]
  (debug "qs" qs)
  (page
   [:h2 "QA: Questions"]
+  [:p "質問をクリックしたら回答ページへ飛ぶ。"]
   (into [:ol]
         (for [q qs]
-          [:li (str (:q q) (:ts q))]))))
+          [:li (str (ss 20 (:q q))
+                    " by " (:nick q)
+                    " at " (st (:ts q)))
+               [:a {:href (str "/as/" (:id q))} " 👉"]]))
+  [:p [:a {:href "/q" :class "btn btn-primary btn-sm"} "new"]]))